+++ pkg/gatewayserver/config.go
@@ -15,6 +15,8 @@
 package gatewayserver
 
 import (
+	"time"
+
 	"go.thethings.network/lorawan-stack/pkg/config"
 	"go.thethings.network/lorawan-stack/pkg/gatewayserver/io/udp"
 	"go.thethings.network/lorawan-stack/pkg/types"
@@ -28,10 +30,11 @@ type UDPConfig struct {
 
 // BasicStationConfig defines the Basic Station configuration of the Gateway Server.
 type BasicStationConfig struct {
-	FallbackFrequencyPlanID string `name:"fallback-frequency-plan-id" description:"Fallback frequency plan ID for non-registered gateways"`
-	Listen                  string `name:"listen" description:"Address for the Basic Station frontend to listen on"`
-	ListenTLS               string `name:"listen-tls" description:"Address for the Basic Station frontend to listen on (with TLS)"`
-	UseTrafficTLSAddress    bool   `name:"use-traffic-tls-address" description:"Use WSS for the traffic address regardless of the TLS setting"`
+	FallbackFrequencyPlanID string        `name:"fallback-frequency-plan-id" description:"Fallback frequency plan ID for non-registered gateways"`
+	Listen                  string        `name:"listen" description:"Address for the Basic Station frontend to listen on"`
+	ListenTLS               string        `name:"listen-tls" description:"Address for the Basic Station frontend to listen on (with TLS)"`
+	UseTrafficTLSAddress    bool          `name:"use-traffic-tls-address" description:"Use WSS for the traffic address regardless of the TLS setting"`
+	WSPingInterval          time.Duration `name:"ws-ping-interval" description:"Interval to send WS ping messages"`
 }
 
 // Config represents the Gateway Server configuration.