+++ pkg/gatewayserver/gatewayserver_test.go
@@ -60,7 +60,8 @@ var (
 	unregisteredGatewayID  = "eui-bbff000000000000"
 	unregisteredGatewayEUI = types.EUI64{0xBB, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00}
 
-	timeout = (1 << 5) * test.Delay
+	timeout        = (1 << 5) * test.Delay
+	wsPingInterval = (1 << 3) * test.Delay
 )
 
 func TestGatewayServer(t *testing.T) {
@@ -102,7 +103,8 @@ func TestGatewayServer(t *testing.T) {
 			},
 		},
 		BasicStation: gatewayserver.BasicStationConfig{
-			Listen: ":1887",
+			Listen:         ":1887",
+			WSPingInterval: wsPingInterval,
 		},
 	}
 	gs, err := gatewayserver.New(c, config)