+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Class B support.
+- WebSocket Ping-Pong support for Basic Station frontend in the Gateway Server.
 
 ### Changed